Adds files chart_generator.py and interest_calc.py.
chart_generator.py: creates a chart of interest rates and candy left with resulting time to breakdown

interest_calc.py: Most important file in the PR.
Contains method that calculates interest rate based on some assumptions that should be easy to calculate in the website.
Can integrate this method into the website later using this data.